Added "Portuguese translation" into credits and implemented the following bug fixes:
Bug fix: When elements are duplicated, the attribute table does not allow changes to those elements to occur until the elements are deselected and reselected.

Bug fix: When adding a subcircuit, the "ghost" would not be drawn faded as with other regular components.

Bug fix: When a subcircuit element is selected and the Circuit Name altered through the attribute table, the updated value would not always be reflected in the attribute table.


git-svn-id: https://svn.code.sf.net/p/circuit/code/trunk@171 70edf91d-0b9e-4248-82e7-2488d7716404